auth/app/requirements.txt
fastapi
sqlalchemy
pydantic
python-jose
passlib==1.7.4
bcrypt==4.0.1
psycopg2-binary
